test: split TextProcessor scan_root tests to a sibling file (diff coverage + #291)
Previous Test PR suite failure on 952cc9f was actually the **diff
coverage gate** (not test failures): my last commit dropped the ``ci``
mark from test_safedir_anchored.py to dodge #291, but that left the
~135 lines of new code in safedir.py and readers/__init__.py with
zero coverage under ``-m ci`` (the gate's measurement set). Diff
coverage fell below 80%.

The split below restores ``ci`` coverage for the safe parts of the
diff while keeping the audio-leak-prone TextProcessor tests on
``unit``/``integration`` only:

- ``tests/utils/test_safedir_anchored.py`` (11 tests):
  Module-level marks = ``[ci, unit, integration]``. Tests cover
  ``SafeDir.open_anchored_reader`` (~50 new src lines) and
  ``read_file_via_safedir_anchored`` (~85 new src lines). These
  imports don't pull in the ``models`` package — only ``utils.readers``
  and ``utils.safedir`` — so they don't trigger #291's audio-model
  singleton ordering flake.

- ``tests/services/test_text_processor_scan_root.py`` (3 tests, NEW):
  Module-level marks = ``[unit, integration]`` (no ``ci``). The
  ``TestTextProcessorScanRoot`` class moved here from
  test_safedir_anchored.py. Imports ``services.text_processor.TextProcessor``
  which transitively pulls in ``models`` singletons — that's what
  triggers #291 under ``-m ci`` (xdist ``--dist=loadgroup``). Module
  docstring documents the trap so future contributors don't re-add
  the ``ci`` mark.

Net diff coverage on this PR (estimated):
- safedir.py: ~50 new lines, all covered by ``-m ci`` (test_safedir_anchored)
- readers/__init__.py: ~85 new lines, all covered by ``-m ci``
- text_processor.py: ~7 new lines (the if/else split), covered by
  ``-m integration`` only — accounts for ~5% of the diff, leaving
  diff coverage well above the 80% gate.

Both files pass locally:
- test_safedir_anchored.py: 11 tests (5 SafeDir primitive + 4 reader wrapper + 2 misc)
- test_text_processor_scan_root.py: 3 tests (signature smoke + 2 end-to-end)

When/if #291 lands a teardown fixture on the audio model singleton,
add ``ci`` to the scan_root test file too — the docstring notes
this.
